fix error String.length value for native string

+++ src/jsvar.c
@@ -1001,7 +1001,7 @@ JsVar *jsvNewNativeString(char *ptr, size_t len) {
   JsVar *str = jsvNewWithFlags(JSV_NATIVE_STRING);
   if (!str) return 0;
   str->varData.nativeStr.ptr = ptr;
-  str->varData.nativeStr.len = (uint16_t)len;
+  str->varData.nativeStr.len = len;
   return str;
 }
